Refactor: extract common-api project
