+++ check-digit/build.sbt
@@ -1,0 +1,1 @@
+
