+++ check-digit/src/main/resources/application.conf
@@ -1,0 +1,12 @@
+akka {
+  log-level = INFO
+}
+
+hmda {
+  uli {
+    http {
+      host = "0.0.0.0"
+      port = 9091
+    }
+  }
+}
