+++ check-digit/src/main/scala/hmda/uli/api/http/HmdaUliApi.scala
@@ -1,0 +1,7 @@
+package hmda.uli.api.http
+
+import akka.actor.Actor
+
+class HmdaUliApi extends Actor {
+  override def receive: Receive = ???
+}
