+++ check-digit/src/test/resources/application.conf
@@ -1,0 +1,4 @@
+akka {
+  stdout-loglevel = OFF
+  loglevel = OFF
+}
