+++ common-api/src/main/scala/hmda/api/http/directives/HmdaTimeDirectives.scala
@@ -1,4 +1,4 @@
-package hmda.api.http.model.directives
+package hmda.api.http.directives
 
 import akka.event.LoggingAdapter
 import akka.http.scaladsl.server.Directive0
