+++ common-api/src/main/scala/hmda/api/http/model/HmdaServiceStatus.scala
@@ -1,4 +1,4 @@
-package hmda.api.http.model.common
+package hmda.api.http.model
 
 final case class HmdaServiceStatus(status: String,
                                    service: String,
