+++ common-api/src/main/scala/hmda/api/http/routes/BaseHttpApi.scala
@@ -1,4 +1,4 @@
-package hmda.api.http
+package hmda.api.http.routes
 
 import java.net.InetAddress
 import java.time.Instant
@@ -12,8 +12,8 @@ import akka.http.scaladsl.server.Directives.{
   pathSingleSlash
 }
 import akka.stream.ActorMaterializer
-import hmda.api.http.model.common.HmdaServiceStatus
-import hmda.api.http.model.directives.HmdaTimeDirectives
+import hmda.api.http.directives.HmdaTimeDirectives
+import hmda.api.http.model.HmdaServiceStatus
 import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport._
 import io.circe.generic.auto._
