+++ common-api/src/main/scala/hmda/messages/CommonMessages.scala
@@ -1,4 +1,4 @@
-package hmda.model.messages
+package hmda.messages
 
 object CommonMessages {
   sealed trait Message
