+++ common-api/src/test/scala/hmda/api/http/routes/BaseHttpApiSpec.scala
@@ -1,14 +1,14 @@
-package hmda.api.http
+package hmda.api.http.routes
 
 import akka.event.{LoggingAdapter, NoLogging}
 import akka.http.scaladsl.model.StatusCodes
-import akka.http.scaladsl.model.headers.{HttpEncodings, `Accept-Encoding`}
 import akka.http.scaladsl.model.headers.HttpEncodings._
+import akka.http.scaladsl.model.headers.{HttpEncodings, `Accept-Encoding`}
 import akka.http.scaladsl.testkit.ScalatestRouteTest
-import hmda.api.http.model.common.HmdaServiceStatus
+import hmda.api.http.model.HmdaServiceStatus
 import org.scalatest.{MustMatchers, WordSpec}
-import io.circe.generic.auto._
 import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport._
+import io.circe.generic.auto._
 
 import scala.concurrent.ExecutionContext
