+++ hmda/src/main/scala/hmda/api/http/HmdaAdminApi.scala
@@ -7,7 +7,7 @@ import akka.http.scaladsl.server.Route
 import akka.pattern.pipe
 import akka.stream.ActorMaterializer
 import com.typesafe.config.ConfigFactory
-import hmda.api.http.model.common.HttpServer
+import hmda.api.http.routes.BaseHttpApi
 
 import scala.concurrent.{ExecutionContext, Future}
